dev(refactor): split to multiple files based on endpoint
